Isolated tests for NACCESS
