+++ Tests/test_PDB.py
@@ -1829,25 +1829,6 @@ class DsspTests(unittest.TestCase):
                     i += 1
 
 
-class NACCESSTests(unittest.TestCase):
-    """Tests for NACCESS parsing etc which don't need the binary tool.
-
-    See also test_NACCESS_tool.py for run time testing with the tool.
-    """
-
-    def test_NACCESS_rsa_file(self):
-        """Test parsing of pregenerated rsa NACCESS file."""
-        with open("PDB/1A8O.rsa") as rsa:
-            naccess = process_rsa_data(rsa)
-        self.assertEqual(len(naccess), 66)
-
-    def test_NACCESS_asa_file(self):
-        """Test parsing of pregenerated asa NACCESS file."""
-        with open("PDB/1A8O.asa") as asa:
-            naccess = process_asa_data(asa)
-        self.assertEqual(len(naccess), 524)
-
-
 class ResidueDepthTests(unittest.TestCase):
     """Tests for ResidueDepth module, except for running MSMS itself."""
